Replaced a typo with rockchip_pmu_set_idle_request
Thanks to DuckDuckGo for the suggestion when I was searching the
mistyped symbol.

Well, I think it's this one. It almost has the same name AND the same
signature. However, I currently do not understand what the code does so
I might be wrong.
The module compiles and loads fine with this correction. However, I do
not know if it runs fine.

Signed-off-by: Myy <myy@miouyouyou.fr>